requirements.txt
opencv-python
easyocr
mysql-connector-python
Pillow
scikit-image
scikit-learn
tensorflow==1.*
matplotlib
numpy